Rename download.py to avoid name conflicts
