+++ src/noisepy/seis/__init__.py
@@ -5,7 +5,8 @@
 import logging
 
 from ._version import __version__  # noqa: F401
-from .download import download  # noqa: F401
+
+from .fdsn_download import download  # noqa: F401
 from .correlate import cross_correlate  # noqa: F401
 from .stack import stack_cross_correlations  # noqa: F401
